requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
aiohttp-cors==0.7.0
python-dotenv==1.0.0
